requirements.txt
jieba
flask
fasttext-wheel
requests
pandas==2.2.3
scikit-learn==1.7.2
icecream
numpy==1.26.4
tqdm
mysql-connector-python
streamlit
langchain_openai
torch==2.4.1
transformers
gensim
matplotlib
torchviz
graphviz
openai
python-dotenv
ollama
langchain
langchain_community
imbalanced-learn
setuptools==67.6.1